+++ R/ohsome_extract_elements.R
@@ -41,7 +41,7 @@
 ohsome_extract_elements <- function(
 	boundary = NULL,
 	geometryType = c("centroid", "bbox", "geometry"),
-	time = lubridate::format_ISO8601(ohsome_temporalExtent[2]),
+	time = ohsome_temporalExtent[2],
 	properties = NULL,
 	clipGeometry = TRUE,
 	...