+++ R/ohsome_extract_elementsFullHistory.R
@@ -29,7 +29,7 @@
 ohsome_extract_elementsFullHistory <- function(
 	boundary = NULL,
 	geometryType = c("centroid", "bbox", "geometry"),
-	time = lubridate::format_ISO8601(ohsome_temporalExtent),
+	time = ohsome_temporalExtent,
 	properties = NULL,
 	clipGeometry = TRUE,
 	...