+++ tests/testthat/test-ohsome_extract_contributions.R
@@ -20,6 +20,6 @@ test_that("correctly sets clipGeometry parameter", {
 test_that("uses full temporal extent if time parameter is missing", {
 	expect_equal(
 		ohsome_extract_contributions()$body$time,
-		paste(lubridate::format_ISO8601(ohsome_temporalExtent), collapse = ",")
+		paste(ohsome_temporalExtent, collapse = ",")
 	)
 })